docs: move status image to child page, float it
The Linux statuses image is a neat idea, but it doesn't belong on
the front page. It's appropriate for the other-installations page.
The problem, though: it's massively vertical, and that's not
changeable. Rather than letting it eat up a lot of vertical space
on the page, what if we float it on the right-hand side of the page?
This isn't a standard approach, but this image is better (IMO) as a
flowed element than a full-page element.