feat(plugins): Support CLI plugins
In the spirit of an iterative and boring solution this commit adds
support for a binary plugin system in the style of `git` and `kubectl`.
To use a plugin it must meet the following two conditions:

1. It must be prefixed with `glab-`.
2. It must be on the user's `$PATH`.

If a command is not found and it's not a shell alias, the CLI will
attempt to look for a binary that matches that matches the pattern
`glab-<ARG1>`. If found, it will run the command and pass all remaining
args to it.

The previous implementation had a high degree of complexity, so the
functionality to run a command and handle any errors has been abstracted
into separate functions.